Feat: 리액트 라우터 기능 추가 및 파일 분리
1. Main 컴포넌트로 리액트 라우터 페이지 분리
2. Shared 페이지 폴더 및 css 파일 분리
