+++ src/components/App/App.jsx
@@ -1,0 +1,15 @@
+import Navbar from "../Navbar/Navbar";
+import Footer from "../Footer/Footer";
+import { Outlet } from "react-router-dom";
+import "../../styles/global.css";
+function App() {
+  return (
+    <>
+      <Navbar />
+      <Outlet />
+      <Footer />
+    </>
+  );
+}
+
+export default App;
