+++ src/components/Footer/Footer.css
@@ -1,0 +1,61 @@
+footer {
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  height: 16rem;
+  padding-top: 3.2rem;
+  background-color: #111322;
+}
+
+.footer-box {
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+  max-width: 192rem;
+  height: fit-content;
+  padding: 0 10.4rem;
+}
+
+.copyright {
+  color: #676767;
+  font-family: Arial;
+  font-size: 1.6rem;
+}
+
+.footer-links {
+  display: flex;
+  column-gap: 3rem;
+  padding-right: 1.8rem;
+}
+
+.footer-link {
+  color: #cfcfcf;
+  font-family: Arial;
+  font-size: 1.6rem;
+}
+
+.sns {
+  display: flex;
+  column-gap: 1.2rem;
+  height: 2rem;
+}
+
+@media (max-width: 767px) {
+  .footer-box {
+    display: grid;
+    row-gap: 6rem;
+    grid-template-areas:
+      "footer-links sns"
+      "copyright .";
+    padding: 0 3.2rem;
+  }
+  .copyright {
+    grid-area: copyright;
+  }
+  .footer-links {
+    grid-area: footer-links;
+  }
+  .sns {
+    grid-area: sns;
+  }
+}
