+++ src/components/Navbar/Navbar.css
@@ -1,0 +1,80 @@
+nav {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  max-width: 192rem;
+  height: 9.4rem;
+  padding: 0 20rem;
+}
+
+.nav-background {
+  position: fixed;
+  top: 0;
+  background-color: #edf7ff;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 2;
+}
+.user-info {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 0.6rem;
+  font-size: 14px;
+  color: var(--gray-50-color);
+}
+nav .profile {
+  width: 2.8rem;
+  border-radius: 70%;
+}
+
+.cta {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 5.4rem;
+  cursor: pointer;
+  background-image: var(--gra-purpleblue-to-skyblue);
+  border-radius: 0.8rem;
+  color: #f5f5f5;
+  font-size: 1.8rem;
+  font-weight: 600;
+}
+
+.cta-short {
+  width: 12.8rem;
+}
+
+@media (max-width: 1124px) {
+  nav {
+    padding: 0;
+    margin: 0 3.2rem;
+    max-width: 70.4rem;
+  }
+  nav .logo-mobile {
+    width: 88.67px;
+    height: 16px;
+  }
+}
+
+@media (max-width: 767px) {
+  nav {
+    height: 6.3rem;
+    max-width: 32.5rem;
+  }
+  .user-info .email {
+    display: none;
+  }
+
+  .cta {
+    height: 3.7rem;
+    font-size: 1.4rem;
+  }
+
+  .cta-short {
+    width: 8rem;
+  }
+}
