+++ src/components/Navbar/Navbar.jsx
@@ -1,7 +1,29 @@
+import { useState, useEffect, useCallback } from "react";
 import logo from "../../assets/logo.svg";
-
-const Navbar = ({ loginClick, isLogin, loginInfo }) => {
+import useAsync from "../../hooks/useAsync";
+import "./Navbar.css";
+import { getApiItems } from "../../services/api";
+const Navbar = () => {
+  const [isLogin, setIsLogin] = useState(false);
+  const [loginInfo, setLoginInfo] = useState([]);
+  const [isLoading, loadingError, getApiItemsAsync] = useAsync(getApiItems);
   const { email, profileImageSource } = loginInfo;
+
+  const handleLoginStatusInfo = useCallback(async () => {
+    let result = await getApiItemsAsync("sample/user");
+    if (!result) return;
+
+    const userLoginInfo = result;
+    if (userLoginInfo) {
+      setIsLogin(true);
+      setLoginInfo(userLoginInfo);
+    }
+  }, [getApiItemsAsync]);
+
+  useEffect(() => {
+    handleLoginStatusInfo();
+  }, [handleLoginStatusInfo]);
+
   return (
     <div className="nav-background">
       <nav>
@@ -12,10 +34,9 @@ const Navbar = ({ loginClick, isLogin, loginInfo }) => {
             <p className="email">{email}</p>
           </div>
         ) : (
-          <button className="cta cta-short" onClick={loginClick}>
-            로그인
-          </button>
+          <button className="cta cta-short">로그인</button>
         )}
+        {loadingError?.message && <span>{loadingError.message}</span>}
       </nav>
     </div>
   );
