+++ src/components/Profile/Profile.css
@@ -1,0 +1,37 @@
+.folder-info {
+  width: 100%;
+  background-color: #edf7ff;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 1.2rem;
+  padding-top: 2rem;
+  padding-bottom: 6rem;
+  margin-top: 9.4rem;
+}
+
+.folder-info .folder-profile {
+  width: 6rem;
+  border-radius: 70%;
+}
+.folder-info #folder-owner-name {
+  font-size: 1.6rem;
+  line-height: 150%;
+}
+.folder-info #folder-name {
+  font-size: 4rem;
+  font-weight: 600;
+}
+
+@media (max-width: 767px) {
+  .folder-info .folder-profile {
+    width: 4rem;
+  }
+  #folder-owner-name {
+    font-size: 1.4rem;
+  }
+  #folder-name {
+    font-size: 3.2rem;
+  }
+}
