+++ src/components/SearchBar/SearchBar.css
@@ -1,0 +1,27 @@
+input {
+  width: 100%;
+  height: 6rem;
+  padding: 1.5rem 1.6rem 1.5rem 4.2rem;
+  border-radius: 8px;
+  border: 1px solid var(--gray-20-color);
+  background: var(--white-color);
+}
+.input-wrapper {
+  position: relative;
+  width: 100%;
+}
+input::placeholder {
+  font-size: 1.6rem;
+  font-weight: 400;
+  color: #666;
+  line-height: 150%;
+}
+
+.input-wrapper #search-icon {
+  position: absolute;
+  transform: translateY(-50%);
+  top: 50%;
+  left: 1.6rem;
+  width: 1.6rem;
+  height: 1.6rem;
+}
