--- src/styles/index.css
@@ -1,310 +1,0 @@
-@import "./global.css";
-
-header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  background-color: #edf7ff;
-}
-
-nav {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  max-width: 192rem;
-  height: 9.4rem;
-  padding: 0 20rem;
-}
-
-.nav-background {
-  position: fixed;
-  background-color: #edf7ff;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 2;
-}
-.user-info {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 0.6rem;
-  font-size: 14px;
-  color: var(--gray-50-color);
-}
-nav .profile {
-  width: 2.8rem;
-  border-radius: 70%;
-}
-
-header .folder-info {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 1.2rem;
-  padding-top: 2rem;
-  padding-bottom: 6rem;
-  margin-top: 9.4rem;
-}
-
-.folder-info .folder-profile {
-  width: 6rem;
-  border-radius: 70%;
-}
-.folder-info #folder-owner-name {
-  font-size: 1.6rem;
-  line-height: 150%;
-}
-.folder-info #folder-name {
-  font-size: 4rem;
-  font-weight: 600;
-}
-.cta {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 5.4rem;
-  cursor: pointer;
-  background-image: var(--gra-purpleblue-to-skyblue);
-  border-radius: 0.8rem;
-  color: #f5f5f5;
-  font-size: 1.8rem;
-  font-weight: 600;
-}
-
-.cta-short {
-  width: 12.8rem;
-}
-
-.cta-long {
-  width: 35rem;
-}
-main .wrapper {
-  margin: 4rem auto 10rem;
-  max-width: 106rem;
-  display: flex;
-  flex-direction: column;
-  gap: 4rem;
-}
-input {
-  width: 100%;
-  height: 6rem;
-  padding: 1.5rem 1.6rem 1.5rem 4.2rem;
-  border-radius: 8px;
-  border: 1px solid var(--gray-20-color);
-  background: var(--white-color);
-}
-.input-wrapper {
-  position: relative;
-  width: 100%;
-}
-input::placeholder {
-  font-size: 1.6rem;
-  font-weight: 400;
-  color: #666;
-  line-height: 150%;
-}
-
-.input-wrapper #search-icon {
-  position: absolute;
-  transform: translateY(-50%);
-  top: 50%;
-  left: 1.6rem;
-  width: 1.6rem;
-  height: 1.6rem;
-}
-
-.card-layout {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-auto-rows: 33.4rem;
-  row-gap: 2.5rem;
-  column-gap: 2rem;
-  width: 100%;
-}
-
-.card {
-  width: 34rem;
-  height: 33.4rem;
-  border-radius: 15px;
-  overflow: hidden;
-  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
-}
-.card-image-box {
-  width: 100%;
-  height: 20rem;
-  overflow: hidden;
-  background-color: #dddfff;
-  position: relative;
-}
-
-.card-image {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  transform: scale(1);
-  transition: transform 0.5s;
-}
-
-.card-image:hover {
-  transform: scale(1.3);
-  transition: transform 0.5s;
-}
-.no-card-image {
-  width: 133px;
-  position: absolute;
-  transform: translate(50%, -50%);
-  top: 50%;
-  right: 50%;
-}
-
-.card-caption {
-  padding: 1.5rem 2rem;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.card-caption .time-msg {
-  font-size: 1.3rem;
-  color: #666;
-}
-.card-caption .description {
-  font-size: 1.6rem;
-  line-height: 150%;
-  height: 4.9rem;
-  text-overflow: ellipsis;
-  overflow: hidden;
-}
-.card-caption .date-msg {
-  font-size: 1.4rem;
-  color: #333;
-  height: 1.9rem;
-}
-
-footer {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  height: 16rem;
-  padding-top: 3.2rem;
-  background-color: #111322;
-}
-
-.footer-box {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-  max-width: 192rem;
-  height: fit-content;
-  padding: 0 10.4rem;
-}
-
-.copyright {
-  color: #676767;
-  font-family: Arial;
-  font-size: 1.6rem;
-}
-
-.footer-links {
-  display: flex;
-  column-gap: 3rem;
-  padding-right: 1.8rem;
-}
-
-.footer-link {
-  color: #cfcfcf;
-  font-family: Arial;
-  font-size: 1.6rem;
-}
-
-.sns {
-  display: flex;
-  column-gap: 1.2rem;
-  height: 2rem;
-}
-
-@media (max-width: 1124px) {
-  nav {
-    padding: 0;
-    margin: 0 3.2rem;
-  }
-  nav .logo-mobile {
-    width: 88.67px;
-    height: 16px;
-  }
-  nav,
-  main .wrapper {
-    max-width: 70.4rem;
-  }
-  .card-layout {
-    grid-template-columns: repeat(2, 1fr);
-    grid-auto-rows: 33.4rem;
-    column-gap: 2.4rem;
-  }
-}
-@media (max-width: 767px) {
-  nav {
-    height: 6.3rem;
-    max-width: 32.5rem;
-  }
-  .user-info .email {
-    display: none;
-  }
-  .cta {
-    height: 3.7rem;
-    font-size: 1.4rem;
-  }
-
-  .cta-short {
-    width: 8rem;
-  }
-  .cta-long {
-    width: 20rem;
-  }
-  .folder-info .folder-profile {
-    width: 4rem;
-  }
-  #folder-owner-name {
-    font-size: 1.4rem;
-  }
-  #folder-name {
-    font-size: 3.2rem;
-  }
-  main .wrapper {
-    margin: 2rem auto 6rem;
-    max-width: 32.5rem;
-    gap: 3.2rem;
-  }
-  .card-layout {
-    grid-template-columns: repeat(1, 1fr);
-    grid-auto-rows: 32.7rem;
-    row-gap: 2rem;
-  }
-  .card {
-    width: 100%;
-    height: 32.7rem;
-  }
-
-  .footer-box {
-    display: grid;
-    row-gap: 6rem;
-    grid-template-areas:
-      "footer-links sns"
-      "copyright .";
-    padding: 0 3.2rem;
-  }
-  .copyright {
-    grid-area: copyright;
-  }
-  .footer-links {
-    grid-area: footer-links;
-  }
-  .sns {
-    grid-area: sns;
-  }
-}
